requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.5
chromadb==0.4.22
pypdf2==3.0.1